+++ app/__init__.py
@@ -7,4 +7,4 @@
 
 __all__ = ["__version__"]
 
-__version__ = "0.1.8"
+__version__ = "0.1.9"